#48 Separate interfaces from the implementations - Separate IReleaseArtifactService from FsReleaseArtifactService - Separate IReleaseArtifactRepository from FsReleaseArtifactRepository - Store them all in separate files
